% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeops.R
\name{scrapeops_fake_useragent}
\alias{scrapeops_fake_useragent}
\title{Query Fake User Agent from ScrapeOps.io}
\usage{
scrapeops_fake_useragent(api_key = Sys.getenv("SCRAPEOPS_API_KEY"))
}
\arguments{
\item{api_key}{ScrapeOps API key, defaults to env var \code{SCRAPEOPS_API_KEY}}
}
\value{
ScrapeOps API returns 10 fake user agents. The function choses one
of them randomly and returns it.
}
\description{
Query Fake User Agent from ScrapeOps.io
}
