% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{.request}
\alias{.request}
\title{base request function}
\usage{
.request(
  url,
  ...,
  verb = "GET",
  retry_delay = 1L,
  retry_times = 3L,
  log_url = url,
  verbose = getOption("undercover.verbose", FALSE)
)
}
\arguments{
\item{url}{Full URL to be scraped.}

\item{...}{arguments to pass to httr request function}

\item{verb}{character: a curl verb supported by httr, e.g. GET, POST, PUT, DELETE
this is passed to \code{httr::RETRY()}. Defaults to \code{"GET"}}

\item{retry_times}{Maximum number of attempts. Defaults to \code{3}.}

\item{log_url}{url for print logging, defaults to \code{url}}

\item{verbose}{whether to print out URLs, default = FALSE}
}
\description{
Formats http request with RETRY and customizable logging
}
